services/chat-service/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.21
psycopg2-binary>=2.9.9
httpx>=0.27.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9
